Add dns server
The access point example is a bit complicated to use as you have to open
a browser and enter a url with the IP address of the device

Add a simple dns server that makes the access point behave like a
captive portal. All DNS requests point back to the gateway IP. The
access point test redirects all web requests.

So now when you connect to the pico AP on a phone it just opens the web
page for you.