DHCP fails sometimes for iperf freertos example
For freertos "sys" examples, timeouts are handled in the tcpip thread.
The default priority of the tcpip thread is 1.

The iperf example creates a thread of priority 2 and this runs
continually until it gets an IP address. So the tcp thread does not
get to run while waiting for an IP address and so DHCP requests are
not re-transmitted when needed.

Raise the priority of the tcp thread to something "highish"

Fixes #296