Make iperf example work with an RM2 breakout board
To use a an RM2 breakout board with a "normal" pico you have to tell
cmake that it supports cyw43 and you have to add the default "pins" to
use to your project.

When building the project pass -DPICO_CYW43_SUPPORTED=1 on the cmake
command line.

cmake -D .. -DPICO_CYW43_SUPPORTED=1

In your project add pin definitions as in this change where WL_ON=2,
CS=3, DAT=4 (in out and wake) and CLK=5. For Pico you seem to have to
slow the pio clock down as well so set div int to 3.

This will only work with the latest SDK 2.1.0.